requirements.txt
keras
Pillow
numpy
pandas
streamlit